\name{pbi-package} 
\alias{pbi-package} 
\title{pbi package - package template for R packages} 
\description{The pbi package can be used as a template to create new packages from scratch.} 
\details{Some more details: 
The following list of objects and/or functions are available: 
\describe{ 
\item{\link[pbi:pbi]{pbi}}{The pbi environment} 
} 
} 
\examples{ 
library(pbi) 
ls(pbi) 
} 
\author{Detlef Groth <dgroth@uni-potsdam.de>} 
\references{ 
\itemize{ 
   \item XYZ  
} 
} 
