\name{pbi_tkregex} 
\alias{pbi$tkregex} 
\alias{pbi_tkregex} 
\title{Graphical user interface for testing regular expressions} 
\description{ 
  This function provides a graphical user interface to execute regular 
  expressions either as PROSITE patterns or a standard regular expressions 
  on the text entered or pasted into the text area. Just copy and paste  
  your text into the text area below of the two entry fields.  
  In the entry fields you can write your regular expression either as normal regular 
  expression in the bottom field or as PROSITE pattern in the top entry. 
  Thereafter press ENTER, your found patterns will be highlighted in red and  
  the number of found patterns will be shown as well. 
} 
\usage{pbi_tkregex()} 
\examples{ 
  \dontrun{  
    pbi_tkregex() 
  } 
} 
\seealso{  
\code{\link[pbi:pbi-package]{pbi-package}}, \code{\link[pbi:pbi-class]{pbi-class}} 
} 
